delegate stream initialisation to another function
